发贴成功👏

+++ src/settings.h
@@ -14,7 +14,7 @@ public:
     Q_INVOKABLE QString get_username();
     Q_INVOKABLE void set_password(const QString &password);
     Q_INVOKABLE QString get_password();
-    Q_INVOKABLE void set_pagesize(const quint16 &pagesize);
+    Q_INVOKABLE void set_pagesize(const int &pagesize);
     Q_INVOKABLE int get_pagesize();
 
 };
